+++ packages/anchor_ui/lib/src/tooltip.dart
@@ -3,7 +3,7 @@ import 'dart:async';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_anchor/flutter_anchor.dart';
 
-import 'arrow/rendering/arrows.dart';
+import 'arrow/arrows.dart';
 import 'popover.dart';
 
 /// Creates an anchor overlay styled as a tooltip.