+++ packages/anchor_ui/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.2
+
+* Better arrow alignment calculation
+
 ## 0.0.1
 
 * Initial release.